fixup: unix-second migration name, tombstone file split, vi allowlist
