+++ backend/internal/modules/privacy/erasure.go
@@ -250,35 +250,6 @@ func purgeRedactedActivityTraces(ctx context.Context, tx pgx.Tx, activities []id
 	return redactDeliveries(ctx, tx, activities, erasedName)
 }
 
-// personErasureCounts is what an erasure tombstone certifies: how much each arm
-// of the cascade removed, and nothing whatsoever about whom.
-type personErasureCounts struct {
-	emailsSuppressed            int
-	rawRowsPurged               int64
-	aiPayloadsPurged            int64
-	activitiesRedacted          int
-	activitiesRestricted        int
-	channelIdentitiesSuppressed int
-}
-
-// tombstonePersonErasure closes the cascade with action=erase and counts only —
-// proof without PII. The counts are evidence ABOUT the scrub, so they ride the
-// evidence column; before/after stay empty — they are reserved for field
-// images, and the record-history read serves a tombstone's images verbatim. The
-// paired event tells consumers the subject is gone.
-func tombstonePersonErasure(ctx context.Context, tx pgx.Tx, subject ids.PersonID, reason string, counts personErasureCounts) error {
-	auditID, err := storekit.AuditWithEvidence(ctx, tx, actionErase, "person", subject.UUID, nil, nil, map[string]any{
-		"reason": reason, "emails_suppressed": counts.emailsSuppressed, "raw_rows_purged": counts.rawRowsPurged,
-		"ai_payloads_purged": counts.aiPayloadsPurged, "activities_redacted": counts.activitiesRedacted,
-		"activities_restricted":         counts.activitiesRestricted,
-		"channel_identities_suppressed": counts.channelIdentitiesSuppressed,
-	})
-	if err != nil {
-		return err
-	}
-	return storekit.EmitEventForEntity(ctx, tx, auditID, "person", subject.UUID, retentionAppliedPayload(actionErase, nil, &reason))
-}
-
 // anonymizeSubjectRows wipes the subject's PII in place: the person row
 // keeps its skeleton (business records other subjects appear in still
 // reference it), the email/phone child rows and the preference-center
